Simplify outerxml, and fix value setting
Outerxjr was previously just displaying the opening and closing tag of the node and the 'contents'. This was missing node attributes. Fixed this by calling the new outerxml function provided by libxjr.

Corrected direct value setting. It was not properly duplicating the key nor the value. ( duplicating as in memory allocating ) Note that this code is still "wrong" in that it does not set the DYN_KEY and DYN_VALUE flags on the node ( which it should ) TODO.