Add new clone function for nodes
Added a new clone function for nodes, and properly set the dynamic name and value flags on a node when setting a new value with memory duplicates name and value.

TODO: Add a unit test to ensure the clone function works as expected.

Also fixed a minor bug in the sethash function. Unsure the exact effect of that bug, but the code was wrong and is more correct now.