Bugfixes
Small bugfixes that were never previously commited that are needed for base functionality. Also changes the magic a bit to use the normal macros instead of the "optimized" shortcut macro. The shortcut macro is fast and skips checks but could break in different perl versions. It seemed to be breaking perl 5.22 and these changes fix that.